requirements.txt
PyMuPDF
langchain==0.3.25
openai==1.81.0
pandas==2.2.3
python-dotenv==1.1.0
langchain-community
faiss-cpu
tiktoken
langchain-openai